Added an experimental flag to delete all graph ops at the end of the deabstraction pass.
Context: Graph mode compilation currently takes ~2 minutes on a resnet 50
model (forward pass only). One conjecture is that running the compiler
performance pipeline over the graph code might have take non-trivial time, and
that could be minimized if we move up the partition pass to be right after
deabstraction, so that we can slice out the accelerator function at the end of
the mandatory pipeline, and do not feel that function through the performance
pipeline (instead, we optimize the accelerator function at the graph level, say
via grappler).

One way to simulate the compiler performance of the above change is to delete
all graph_ops after deabstraction, so that the performance pipeline only pays
cost in optimizing host code.

Unfortunately, the performance gain is not significant. The time cost of
compiling a resnet model (source code to be released) under -DuseToyModel is
17.6s (with 27900 SIL instructions in the deabstraction result). After deleting
the graph ops via the new flag `-Xllvm -tf-delete-graph-ops`, the time cost
reduced to 14.6s (with 27016 SIL instructions).

This suggests that we should look into other venues for compiler performance
optimization first, such as modular (non-inline-based) compilation.

This patch is to submit the new code path for deleting graph_ops under a flag,
so that this experiment can be performed on other future models in case needed.