Fixed a bug in `TensorFunctionClassifier::shouldBePartitioned()`, which returned true for `_hostOp<A>(_:)` defined in the tensorflow stdlib, when we are compiling a user module.
The bug was because we were evaluating `if (fn->getInlineStrategy() ==
NoInline)` (it is true for `_hostOp<A>(_:)`) before evaluating
`if (isAvailableExternally(fn->getLinkage()))` (also true for
`_hostOp<A>(_:)`). But for any externally defined functions, they should not be
partitioned when compiling the user module.

Rearranged the code and added comments to improve its readability. Now we first
handle cases where we return false, and then the cases where we return true.