allow @differentiable to specify JVP and VJP (#20689)
This PR allows users to specify JVP and VJP in the @differentiable attr, and adds typechecking for it.

This PR leaves @differentiable in a weird state where you can specify primal/adjoint and/or jvp/vjp but where only primal/adjoint are fully plumbed through everything. I think that this is acceptable as a temporary intermediate state for switching from primal/adjoint to jvp/vjp.

Note that having JVP and VJP in the @differentiable attr is throwaway work because our long-term design does away with using @differentiable to specify custom associated functions. This throwaway work will unblock my ability to put the JVP and VJP in the witness table though. Also, the type calculations in this PR should not be throwaway.