[5.0] stdlib: Make ArrayBufferProtocol.init inlinable to facilitate specialization
Change the attribute of init(copying:) from @usableFromInline to
@inlinable to enable generic specialization. Resulting in a 2x speedup
of array copying.

This is not ABI affecting because both @usableFromInline and @inlinable
cause emission of the symbol in the standard library dylib.

This is a fix for a performance regression from a previously shipped
swift and it is not a correctness fix.

SR-9284
rdar://46277030