Re-enable a test that had been disabled at one point.
We saw a failure of this test on Linux at some point, and disabled
it. It looks like it may have been due to the fact that the
`%target-typecheck-verify-swift` provides the filename, but we
explicitly provide the filename in the test invocation here as well.

I'm going to re-enable and see if we can get through tests on all
supported platforms.