Limit access of implicit typealiases for associated type witnesses
When conformance checking infers an associated type, it makes a new
typealias to represent that type. Unfortunately, the access of that
typealias was always the same as the conforming type's, which (1)
might have leaked implementation details if the protocol wasn't as
public as the conforming type, and (2) might have been straight-up
incorrect if the /inferred type/ wasn't as public as the conforming
type.

Fix this in pre-Swift-5 modes by limiting the access to fit the
underlying type (technically source-breaking, but most code that
relied on this would have failed to link anyway), and in Swift 5 mode
by using the minimum possible access (the intersection of the protocol
and conforming type).

rdar://problem/46143405