[benchmark] Eliminate output buffers; collapse run() into run(iterations:); fix memory accumulation
I planned to run some validation on the accumulated output in tearDown, but benchmarks don’t seem the right place for doing that.

Having the run() vs run(iterations:) split interfered with getting e.g. Array.withUnsafeMutableBufferPointer moved out of the outer loop.

MutatedIdxToUTF16 and its dual used to append a space to the input string on every single iteration of the inner loop, accumulating spaces indefinitely, even across the outer iterations. It now works on a per-iteration copy of the input string and alternates between appending/removing a final character.