[IRGen] Emit base protocol witnesses as mangled names.
Rather than setting the witness tables for base protocols in a witness
table’s instantiation function, use mangled names as we do for
associated conformances. The runtime will eagerly realize them so clients
of witness tables are not affected.

For now, only do this for unconditional conformances: conditional
conformances won’t work yet because the instantiation function currently
copies the instantiation arguments (which satisfy the conditional
requirements) in to the private area. The runtime will need to take over
this responsibility for conditional conformances to work.

Part of rdar://problem/46282080.